requirements.txt
manim>=0.19.0
openai>=1.0.0
PyYAML>=6.0
imageio-ffmpeg>=0.4.0
google-generativeai>=0.4.0